Correciones en FrmAutor, FrmGenero y FrmEditorial. Varios mensajes añadidos de info.

+++ OpenLibraryEditor/OpenLibraryEditor/Forms/FrmAutores.cs
@@ -69,6 +69,8 @@ namespace OpenLibraryEditor.Forms
             }
 
             //Vincular lista de ocupacion con Combobox
+            //Limpiar objetos nulos o vacios si se cuela alguno
+            listaOcupacion.RemoveAll(p => String.IsNullOrWhiteSpace(p));
             ocupacionBinding.DataSource = listaOcupacion;
             KCmbOcupacionNA.DataSource = ocupacionBinding;
 
@@ -81,27 +83,27 @@ namespace OpenLibraryEditor.Forms
             {
                 GBtnActualizar.Visible = true;
             }
-            
+            LsvAutoresNA_ItemSelectionChanged(null, null);
         }
         #endregion
         #region metodos propios
-        private void EditarAutorDesdeMain()
-        {
-            PanOpcionesNA.Visible = true;
-            KTxtNombreAu.Text = autorActual.Nombre;
-            KTxtComentarioAu.Text = autorActual.Comentario;
-            KCmbOcupacionNA.SelectedItem = String.IsNullOrWhiteSpace(autorActual.NombreOcupacion)
-                    ?
-                    null : autorActual.NombreOcupacion;
-            KtxtAliasAu.Text = autorActual.Alias;
-            KTxtEnlaceAu.Text = autorActual.EnlaceReferencia;
-            rutaImagen = autorActual.Imagen;
-            CargarImagen(rutaImagen);
-            if (!autorActual.FechaDefuncion.Date.Equals(new DateTime()))
-                KMtxtFecMuerteNA.Text = autorActual.FechaDefuncion.Date.ToShortDateString();
-            if (!autorActual.FechaNacimiento.Date.Equals(new DateTime()))
-                KMtxtFecNacimientoNA.Text = autorActual.FechaNacimiento.Date.ToShortDateString();
-        }
+        //private void EditarAutorDesdeMain()
+        //{
+        //    PanOpcionesNA.Visible = true;
+        //    KTxtNombreAu.Text = autorActual.Nombre;
+        //    KTxtComentarioAu.Text = autorActual.Comentario;
+        //    KCmbOcupacionNA.SelectedItem = String.IsNullOrWhiteSpace(autorActual.NombreOcupacion)
+        //            ?
+        //            null : autorActual.NombreOcupacion;
+        //    KtxtAliasAu.Text = autorActual.Alias;
+        //    KTxtEnlaceAu.Text = autorActual.EnlaceReferencia;
+        //    rutaImagen = autorActual.Imagen;
+        //    CargarImagen(rutaImagen);
+        //    if (!autorActual.FechaDefuncion.Date.Equals(new DateTime()))
+        //        KMtxtFecMuerteNA.Text = autorActual.FechaDefuncion.Date.ToShortDateString();
+        //    if (!autorActual.FechaNacimiento.Date.Equals(new DateTime()))
+        //        KMtxtFecNacimientoNA.Text = autorActual.FechaNacimiento.Date.ToShortDateString();
+        //}
         private void IdiomaTexto()
         {
             LblTituloAutores.Text = ControladorIdioma.GetTexto("Au_TituloFrm");
@@ -158,10 +160,12 @@ namespace OpenLibraryEditor.Forms
 
         private bool EsObjetoCambiado()
         {
-            string a = KMtxtFecMuerteNA.MaskCompleted ? DateTime.Parse(KMtxtFecMuerteNA.Text).Date.ToShortDateString()
+            string a = KMtxtFecMuerteNA.MaskCompleted ? KMtxtFecMuerteNA.Text
                 : new DateTime().Date.ToShortDateString();
-            string b = KMtxtFecNacimientoNA.MaskCompleted ? DateTime.Parse(KMtxtFecNacimientoNA.Text).Date.ToShortDateString()
+            string b = KMtxtFecNacimientoNA.MaskCompleted ? KMtxtFecNacimientoNA.Text
                 : new DateTime().Date.ToShortDateString();
+            string c = autorActual.NombreOcupacion == null ? ""
+                : autorActual.NombreOcupacion;
 
             //Comprobar si el objeto actual tiene algo cambiado
             if (KTxtNombreAu.Text == autorActual.Nombre &&
@@ -169,7 +173,7 @@ namespace OpenLibraryEditor.Forms
                 KtxtAliasAu.Text == autorActual.Alias &&
                 KTxtEnlaceAu.Text == autorActual.EnlaceReferencia &&
                 rutaImagen == autorActual.Imagen &&
-                KCmbOcupacionNA.Text == autorActual.NombreOcupacion &&
+                KCmbOcupacionNA.Text == c &&
                 a == autorActual.FechaDefuncion.Date.ToShortDateString() &&
                 b == autorActual.FechaNacimiento.Date.ToShortDateString())
                 return false;
@@ -218,7 +222,7 @@ namespace OpenLibraryEditor.Forms
                 ComprobarGuardado();
 
             //Comprobar selección item
-            if (e == null || (e.IsSelected && LsvAutoresNA.SelectedItems.Count == 1))
+            if ((e == null || e.IsSelected) && LsvAutoresNA.SelectedItems.Count == 1)
             {
                 PanOpcionesNA.Visible = true;
                 itemActual = LsvAutoresNA.SelectedItems[0];
@@ -226,9 +230,15 @@ namespace OpenLibraryEditor.Forms
                 KTxtNombreAu.Text = autorActual.Nombre;
                 KTxtComentarioAu.Text = autorActual.Comentario;
                 if (String.IsNullOrWhiteSpace(autorActual.NombreOcupacion))
+                {
+                    KCmbOcupacionNA.SelectedItem = null;
                     KCmbOcupacionNA.Text = "";
+                }
                 else
+                {
                     KCmbOcupacionNA.SelectedItem = autorActual.NombreOcupacion;
+                    KCmbOcupacionNA.Text = autorActual.NombreOcupacion;
+                }
                 KtxtAliasAu.Text = autorActual.Alias;
                 KTxtEnlaceAu.Text = autorActual.EnlaceReferencia;
                 rutaImagen = autorActual.Imagen;
@@ -236,11 +246,11 @@ namespace OpenLibraryEditor.Forms
                 if (!autorActual.FechaDefuncion.Date.Equals(new DateTime()))
                 {
                     KMtxtFecMuerteNA.Text = autorActual.FechaDefuncion.Date.ToShortDateString();
-                    TBtnVivo.Checked = true;
+                    TBtnVivo.Checked = false;
                 }
                 else
                 {
-                    TBtnVivo.Checked = false;
+                    TBtnVivo.Checked = true;
                     KMtxtFecMuerteNA.Text = "";
                 }
                 TBtnVivo_CheckedChanged(null, null);
@@ -262,6 +272,7 @@ namespace OpenLibraryEditor.Forms
             listaPersona.Add(p);
             var item = AniadirPersona(p);
             item.Selected = true;
+            Biblioteca.biblioteca.GuardarJson();
         }
 
         private void MBtnMenosLsvNA_Click(object sender, EventArgs e)
@@ -280,6 +291,8 @@ namespace OpenLibraryEditor.Forms
                         {
                             autorActual.BorraDeBDCompartida();
                             ConexionBD.CerrarConexion();
+                            //---------------
+                            VentanaWindowsComun.MensajeInformacion("Se ha borrado correctamente de la base de datos.");
                         }
                     }
                 }
@@ -295,8 +308,12 @@ namespace OpenLibraryEditor.Forms
 
                 listaPersona.Remove(autorActual);
                 LsvAutoresNA.Items.Remove(item);
+                Biblioteca.biblioteca.GuardarJson();
                 VentanaWindowsComun.MensajeInformacion(NOMBRE_OBJETO+ControladorIdioma.GetTexto("BorradoCorrectamente"));
             }
+            else
+                //------------
+                VentanaWindowsComun.MensajeInformacion("Debes seleccionar un autor de la lista para realizar esta operación.");
         }
         private void MBtnMasOcupacionNA_Click(object sender, EventArgs e)
         {
@@ -326,8 +343,19 @@ namespace OpenLibraryEditor.Forms
             else if (x == null)
             {
                 listaOcupacion.Remove((string)KCmbOcupacionNA.SelectedItem);
+                listaPersona.ForEach(p =>
+                {
+                    if (p.NombreOcupacion == KCmbOcupacionNA.Text)
+                        p.NombreOcupacion = "";
+                });
+                foreach (ListViewItem item in LsvAutoresNA.Items)
+                {
+                    if (item.SubItems[1].Text == KCmbOcupacionNA.Text)
+                        item.SubItems[1].Text = "";
+                }
                 ActualizarOcupacion();
                 KCmbOcupacionNA.SelectedItem = null;
+                KCmbOcupacionNA.Text = "";
             }
         }
         private void MBtnAniadirImagenAu_Click(object sender, EventArgs e)
@@ -348,8 +376,8 @@ namespace OpenLibraryEditor.Forms
         private void ActualizarListView()
         {
             //Actualizar listview
-            itemActual.Text = KTxtNombreAu.Text;
-            itemActual.SubItems[1].Text = KCmbOcupacionNA.Text;
+            itemActual.Text = autorActual.Nombre;
+            itemActual.SubItems[1].Text = autorActual.NombreOcupacion;
         }
 
         private void KTxtNombreAu_Enter(object sender, EventArgs e)
@@ -384,7 +412,7 @@ namespace OpenLibraryEditor.Forms
                     autorActual.Comentario = KTxtComentarioAu.Text;
                     autorActual.Alias = KtxtAliasAu.Text;
                     autorActual.EnlaceReferencia = KTxtEnlaceAu.Text;
-                    if (KMtxtFecMuerteNA.MaskCompleted && TBtnVivo.Checked)
+                    if (KMtxtFecMuerteNA.MaskCompleted && !TBtnVivo.Checked)
                     {
                         DateTime d = new DateTime();
                         DateTime.TryParse(KMtxtFecMuerteNA.Text, out d);
@@ -429,58 +457,77 @@ namespace OpenLibraryEditor.Forms
                             {
                                 autorActual.MeterEnBDCompartida();
                                 ConexionBD.CerrarConexion();
+                                //---------------
+                                VentanaWindowsComun.MensajeInformacion("Se ha guardado correctamente en la base de datos.");
                             }
                         }
                     }
 
                     ActualizarListView();
+                    Biblioteca.biblioteca.GuardarJson();
                     VentanaWindowsComun.MensajeInformacion(NOMBRE_OBJETO + ControladorIdioma.GetTexto("GuardadoCorrectamente"));
                 }
                 else
                     VentanaWindowsComun.MensajeError(ControladorIdioma.GetTexto("Error_NombreVacio"));
             }
+            else
+                //------------
+                VentanaWindowsComun.MensajeInformacion("Debes seleccionar un autor de la lista para realizar esta operación.");
         }
 
      
         private void GBtnActualizar_Click(object sender, EventArgs e)
         {
-            //Actualizar en BD compartida si se puede
-            if (ConexionBD.Conexion != null)
+            if (PanOpcionesNA.Visible == true)
             {
-                if (VentanaWindowsComun.MensajePregunta(ControladorIdioma.GetTexto("VWC_ActualizarAutorBD"))
-                    == DialogResult.Yes)
+                //Actualizar en BD compartida si se puede
+                if (ConexionBD.Conexion != null)
                 {
-                    if (ConexionBD.AbrirConexion())
+                    if (VentanaWindowsComun.MensajePregunta(ControladorIdioma.GetTexto("VWC_ActualizarAutorBD"))
+                        == DialogResult.Yes)
                     {
-                        Autor autor = LecturaBD.SelectAutor(EscrituraBD.GetObjetoIdDeLocal(
-                            autorActual.ListaIdCompartido));
-                        ConexionBD.CerrarConexion();
-                        if (autor != null)
+                        if (ConexionBD.AbrirConexion())
                         {
-                            autorActual.Nombre = autor.Nombre;
-                            autorActual.Alias = autor.Alias;
-                            autorActual.Comentario = autor.Comentario;
-                            autorActual.EnlaceReferencia = autor.EnlaceReferencia;
-                            autorActual.FechaDefuncion = autor.FechaDefuncion;
-                            autorActual.FechaNacimiento = autor.FechaNacimiento;
-                            autorActual.Imagen = autor.ImagenTemp == null ? null :
-                                ControladorImagen.RUTA_PERSONA + autorActual.IdAutor;
-                            autorActual.NombreOcupacion = autor.NombreOcupacion;
-                            if (autorActual.Imagen != null)
+                            Autor autor = LecturaBD.SelectAutor(EscrituraBD.GetObjetoIdDeLocal(
+                                autorActual.ListaIdCompartido));
+                            ConexionBD.CerrarConexion();
+                            if (autor != null)
                             {
-                                //Thread.Sleep(10);
-                                File.Delete(autorActual.Imagen);
-                                File.WriteAllBytes(autorActual.Imagen, autor.ImagenTemp);
-                                CargarImagen(autorActual.Imagen);
+                                autorActual.Nombre = autor.Nombre;
+                                autorActual.Alias = autor.Alias;
+                                autorActual.Comentario = autor.Comentario;
+                                autorActual.EnlaceReferencia = autor.EnlaceReferencia;
+                                autorActual.FechaDefuncion = autor.FechaDefuncion;
+                                autorActual.FechaNacimiento = autor.FechaNacimiento;
+                                autorActual.Imagen = autor.ImagenTemp == null ? null :
+                                    ControladorImagen.RUTA_PERSONA + autorActual.IdAutor;
+                                if (!String.IsNullOrWhiteSpace(autor.NombreOcupacion) &&
+                                    !listaOcupacion.Contains(autor.NombreOcupacion))
+                                    listaOcupacion.Add(autor.NombreOcupacion);
+                                autorActual.NombreOcupacion = autor.NombreOcupacion;
+                                if (autorActual.Imagen != null)
+                                {
+                                    //Thread.Sleep(10);
+                                    File.Delete(autorActual.Imagen);
+                                    File.WriteAllBytes(autorActual.Imagen, autor.ImagenTemp);
+                                    CargarImagen(autorActual.Imagen);
+                                }
+                                ActualizarOcupacion();
+                                LsvAutoresNA_ItemSelectionChanged(null, null);
+                                ActualizarListView();
+                                Biblioteca.biblioteca.GuardarJson();
+                                //---------------
+                                VentanaWindowsComun.MensajeInformacion("Se ha descargado correctamente de la base de datos.");
                             }
-                            LsvAutoresNA_ItemSelectionChanged(null, null);
-                            ActualizarListView();
+                            else
+                                VentanaWindowsComun.MensajeError(ControladorIdioma.GetTexto("AutorNoExiste"));
                         }
-                        else
-                            VentanaWindowsComun.MensajeError("AutorNoExiste");
                     }
                 }
             }
+            else
+                //------------
+                VentanaWindowsComun.MensajeInformacion("Debes seleccionar un autor de la lista para realizar esta operación.");
         }
         #endregion
         #region cerrar formulario
@@ -513,8 +560,7 @@ namespace OpenLibraryEditor.Forms
 
         private void TBtnVivo_CheckedChanged(object sender, EventArgs e)
         {
-            KMtxtFecMuerteNA.Enabled = TBtnVivo.Checked;
-
+            KMtxtFecMuerteNA.Enabled = !TBtnVivo.Checked;
         }
     }
 }
